Move `CodableFeedStore` to new file in production
